# Scheduling key optimisations