--- a/internal/scheduler/pool_assigner.go
+++ b/internal/scheduler/pool_assigner.go
@@ -29,17 +29,18 @@ type executor struct {
 }
 
 type DefaultPoolAssigner struct {
-	executorTimeout    time.Duration
-	priorityClasses    map[string]configuration.PriorityClass
-	priorities         []int32
-	indexedResources   []string
-	indexedTaints      []string
-	indexedNodeLabels  []string
-	poolByExecutorId   map[string]string
-	executorsByPool    map[string][]*executor
-	executorRepository database.ExecutorRepository
-	poolCache          *lru.Cache
-	clock              clock.Clock
+	executorTimeout        time.Duration
+	priorityClasses        map[string]configuration.PriorityClass
+	priorities             []int32
+	indexedResources       []string
+	indexedTaints          []string
+	indexedNodeLabels      []string
+	poolByExecutorId       map[string]string
+	executorsByPool        map[string][]*executor
+	executorRepository     database.ExecutorRepository
+	schedulingKeyGenerator *schedulerobjects.SchedulingKeyGenerator
+	poolCache              *lru.Cache
+	clock                  clock.Clock
 }
 
 func NewPoolAssigner(executorTimeout time.Duration,
@@ -51,17 +52,18 @@ func NewPoolAssigner(executorTimeout time.Duration,
 		return nil, errors.Wrap(err, "error  creating PoolAssigner pool cache")
 	}
 	return &DefaultPoolAssigner{
-		executorTimeout:    executorTimeout,
-		priorityClasses:    schedulingConfig.Preemption.PriorityClasses,
-		executorsByPool:    map[string][]*executor{},
-		poolByExecutorId:   map[string]string{},
-		priorities:         schedulingConfig.Preemption.AllowedPriorities(),
-		indexedResources:   schedulingConfig.IndexedResources,
-		indexedTaints:      schedulingConfig.IndexedTaints,
-		indexedNodeLabels:  schedulingConfig.IndexedNodeLabels,
-		executorRepository: executorRepository,
-		poolCache:          poolCache,
-		clock:              clock.RealClock{},
+		executorTimeout:        executorTimeout,
+		priorityClasses:        schedulingConfig.Preemption.PriorityClasses,
+		executorsByPool:        map[string][]*executor{},
+		poolByExecutorId:       map[string]string{},
+		priorities:             schedulingConfig.Preemption.AllowedPriorities(),
+		indexedResources:       schedulingConfig.IndexedResources,
+		indexedTaints:          schedulingConfig.IndexedTaints,
+		indexedNodeLabels:      schedulingConfig.IndexedNodeLabels,
+		executorRepository:     executorRepository,
+		schedulingKeyGenerator: schedulerobjects.NewSchedulingKeyGenerator(),
+		poolCache:              poolCache,
+		clock:                  clock.RealClock{},
 	}, nil
 }
 
@@ -88,6 +90,7 @@ func (p *DefaultPoolAssigner) Refresh(ctx context.Context) error {
 	}
 	p.executorsByPool = executorsByPool
 	p.poolByExecutorId = poolByExecutorId
+	p.schedulingKeyGenerator = schedulerobjects.NewSchedulingKeyGenerator()
 	p.poolCache.Purge()
 	return nil
 }
@@ -100,10 +103,19 @@ func (p *DefaultPoolAssigner) AssignPool(j *jobdb.Job) (string, error) {
 	}
 
 	// See if we have this set of reqs cached.
-	if schedulingKey, ok := j.JobSchedulingInfo().SchedulingKey(); ok {
-		if cachedPool, ok := p.poolCache.Get(schedulingKey); ok {
-			return cachedPool.(string), nil
-		}
+	var priority int32
+	if priorityClass, ok := p.priorityClasses[j.GetPriorityClassName()]; ok {
+		priority = priorityClass.Priority
+	}
+	schedulingKey := p.schedulingKeyGenerator.Key(
+		j.GetNodeSelector(),
+		j.GetAffinity(),
+		j.GetTolerations(),
+		j.GetResourceRequirements().Requests,
+		priority,
+	)
+	if cachedPool, ok := p.poolCache.Get(schedulingKey); ok {
+		return cachedPool.(string), nil
 	}
 
 	req := PodRequirementFromJobSchedulingInfo(j.JobSchedulingInfo())
@@ -124,9 +136,7 @@ func (p *DefaultPoolAssigner) AssignPool(j *jobdb.Job) (string, error) {
 					return "", errors.WithMessagef(err, "error selecting node for job %s", j.Id())
 				}
 				if report.Node != nil {
-					if schedulingKey, ok := j.JobSchedulingInfo().SchedulingKey(); ok {
-						p.poolCache.Add(schedulingKey, pool)
-					}
+					p.poolCache.Add(schedulingKey, pool)
 					return pool, nil
 				}
 			}