--- a/internal/scheduler/reports_test.go
+++ b/internal/scheduler/reports_test.go
@@ -264,5 +264,6 @@ func testSchedulingContext(executorId string) *schedulercontext.SchedulingContex
 	)
 	sctx.Started = time.Time{}
 	sctx.Finished = time.Time{}
+	sctx.SchedulingKeyGenerator = nil
 	return sctx
 }