--- a/internal/scheduleringester/instructions.go
+++ b/internal/scheduleringester/instructions.go
@@ -381,9 +381,5 @@ func (c *InstructionConverter) schedulingInfoFromSubmitJob(submitJob *armadaeven
 	default:
 		return nil, errors.Errorf("unsupported object type %T", object)
 	}
-
-	// Call SchedulingKey() to trigger computing and caching the key.
-	_, _ = schedulingInfo.SchedulingKey()
-
 	return schedulingInfo, nil
 }