The current scheduling key implementation is relatively slow and requires many allocs, causing unnecessary GC pressure. Here, I've switched to a new implementation that is much faster and doesn't require allocs. Further, affinities that are equivalent from a scheduling point of view but differ in some other way will now map to the same hash. I've added tests to check that the affinity docs don't change unexpectedly to avoid the affinity definition changing causing us issues.

The new implementation is about 5x faster for jobs without affinities and about 100x faster for jobs with affinities.

The speedup is in large part due to using a 256-bit highway hash instead of sha1. While sha1 is designed to make it hard to find X and Y such that hash(X) = hash(Y), highway hash is designed to make it hard to find Y such that hash(Y) = hash(X, k), where k is an unknown key and X and hash(X) are known. As a result, highway hash can be made more efficient and is a better fit for this application than sha1. For affinities, the speedup is largely due to hashing the relevant parts of the affinities directly instead of serialising to json and hashing the resulting byte array. 

Current implementation:
```
go.exe test -benchmem -run=^$ -bench ^BenchmarkSchedulingKey$ github.com/armadaproject/armada/internal/scheduler/schedulerobjects
goos: windows
goarch: amd64
pkg: github.com/armadaproject/armada/internal/scheduler/schedulerobjects
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkSchedulingKey-20         646980              1796 ns/op             640 B/op         39 allocs/op
PASS
ok      github.com/armadaproject/armada/internal/scheduler/schedulerobjects     1.218s


go.exe test -benchmem -run=^$ -bench ^BenchmarkSchedulingKey_Affinity$ github.com/armadaproject/armada/internal/scheduler/schedulerobjects
goos: windows
goarch: amd64
pkg: github.com/armadaproject/armada/internal/scheduler/schedulerobjects
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkSchedulingKey_Affinity-20         31515             38902 ns/op           17025 B/op        504 allocs/op
PASS
ok      github.com/armadaproject/armada/internal/scheduler/schedulerobjects     1.651s
```

New implementation:
```
go.exe test -v -benchmem -run=^$ -bench ^BenchmarkSchedulingKey$ github.com/armadaproject/armada/internal/scheduler/schedulerobjects
goos: windows
goarch: amd64
pkg: github.com/armadaproject/armada/internal/scheduler/schedulerobjects
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkSchedulingKey
BenchmarkSchedulingKey-20        3922611               304.3 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/armadaproject/armada/internal/scheduler/schedulerobjects     1.540s

go.exe test -v -benchmem -run=^$ -bench ^BenchmarkSchedulingKey_Affinity$ github.com/armadaproject/armada/internal/scheduler/schedulerobjects
goos: windows
goarch: amd64
pkg: github.com/armadaproject/armada/internal/scheduler/schedulerobjects
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
BenchmarkSchedulingKey_Affinity
BenchmarkSchedulingKey_Affinity-20       3030038               394.9 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/armadaproject/armada/internal/scheduler/schedulerobjects     1.636s
```

┆Issue is synchronized with this [Jira Task](https://gr-oss.atlassian.net/browse/BATCH-319) by [Unito](https://www.unito.io)