@severinson 2023-06-07
For some reason the affinity tests aren't showing in the diff. They're here:
https://github.com/armadaproject/armada/blob/e5b8b70aaacac2d666e30f83bc7ab4d073f25286/internal/scheduler/schedulerobjects/podutils_test.go